+++ src/pages/NewsList.tsx
@@ -1,15 +1,17 @@
-import React, { useState, useEffect } from "react"
+import React, { useState, useEffect, useRef } from "react"
 import { View } from "react-native"
 
 import { api, Article, RetryType } from "api"
-import { RootStackScreen } from "navigation/NavigationTypes"
-import { Page } from "components/Page"
+import { RootStackScreen, useNavigation } from "navigation/NavigationTypes"
+import { ScrollPageInfinite } from "components/ScrollPageInfinite"
 import { CardWithGradient } from "components/CardWithGradient"
 
 /**
  * News page containing the articles of a specific category.
  */
 export const NewsList: RootStackScreen<"NewsList"> = props => {
+    const navigation = useNavigation()
+
     // Name of the news category, which is the title of the page
     const { categoryName } = props.route.params
 
@@ -18,57 +20,87 @@ export const NewsList: RootStackScreen<"NewsList"> = props => {
 
     const [articles, setArticles] = useState<Article[]>([])
 
-    const [refreshing, setRefreshing] = useState<boolean>(false)
+    const [manualRefreshing, setManualRefreshing] = useState<boolean>(false)
+    const [automaticDownload, setAutomaticDownload] = useState<boolean>(false)
+
+    const lastDate = useRef<Date>(new Date())
 
     // TODO: get only the articles of the given category
-    const updateArticles = async (retryType: RetryType) => {
+    const updateArticles = async (retryType: RetryType, keep = true) => {
         try {
-            const response = await api.getArticlesFromDaysAgoTillDate(
-                3,
+            const response = await api.getArticlesFromDaysAgoTillDateByTag(
+                7,
                 new Date().toISOString(),
-                { retryType: retryType }
+                categoryName,
+                { retryType: retryType, maxRetries: 2, waitingTime: 2 }
             )
-            setArticles(response)
+
+            if (keep) {
+                setArticles([...articles, ...response.reverse()])
+            } else {
+                setArticles(response.reverse())
+            }
         } catch (error) {
             console.log(error)
         }
-        setRefreshing(false)
+        setManualRefreshing(false)
     }
 
     useEffect(() => {
-        void updateArticles(RetryType.RETRY_INDEFINETELY)
+        void updateArticles(RetryType.RETRY_N_TIMES)
     }, [])
 
     return (
-        <Page
+        <ScrollPageInfinite
             title={categoryName}
+            data={articles}
+            render={article => {
+                return (
+                    <View style={{ paddingHorizontal: 28 }}>
+                        <CardWithGradient
+                            key={article.id} // TODO: update with article id
+                            title={article.title}
+                            imageURL={article.image} // TODO: update with correct article image
+                            onClick={() =>
+                                navigation.navigate("Article", {
+                                    article: article,
+                                })
+                            }
+                            style={{ height: 220, marginBottom: 13 }}
+                        />
+                    </View>
+                )
+            }}
+            fetchData={() => {
+                void updateArticles(RetryType.NO_RETRY, false)
+            }}
             refreshControl={{
-                refreshing,
+                refreshing: manualRefreshing,
                 onRefresh: () => {
-                    setRefreshing(true)
-                    void updateArticles(RetryType.NO_RETRY)
+                    setManualRefreshing(true)
+                    void updateArticles(RetryType.NO_RETRY, false)
                 },
             }}
             showSwitch={true}
             switchValue={toggled}
             onSwitchToggle={value => setToggled(value)}
-        >
-            <View style={{ flex: 1, paddingTop: 16 }}>
-                {articles.map((article, index) => {
-                    return (
-                        <CardWithGradient
-                            key={index} // TODO: update with article id
-                            title={article.title}
-                            imageURL={undefined} // TODO: update with correct article image
-                            onClick={() => {
-                                // TODO: here navigate to the article page
-                                console.log(article.title)
-                            }}
-                            style={{ height: 220, marginBottom: 13 }}
-                        />
-                    )
-                })}
-            </View>
-        </Page>
+        />
+        // {/* <View style={{ flex: 1, paddingTop: 16 }}> ////////
+        //     {articles.map((article, index) => {
+        //         return (
+        //             <CardWithGradient
+        //                 key={index} // TODO: update with article id
+        //                 title={article.title}
+        //                 imageURL={article.image} // TODO: update with correct article image
+        //                 onClick={() =>
+        //                     navigation.navigate("Article", {
+        //                         article: article,
+        //                     })
+        //                 }
+        //                 style={{ height: 220, marginBottom: 13 }}
+        //             />
+        //         )
+        //     })}
+        // </View> */}
     )
 }
